Add understanding of what is imported and track function call versus object level pending tasks and cleaning. Respect overwrite setting as for normal GRASS rasters.

+++ python/grass/tools/importexport.py
@@ -1,16 +1,26 @@
+from __future__ import annotations
+
 import subprocess
 from pathlib import Path
+from typing import Literal
 
 
 class ImporterExporter:
+    """Imports and exports data while keeping track of it
+
+    This is a class for internal use, but it may mature into a generally useful tool.
+    """
+
     raster_pack_suffixes = (".grass_raster", ".pack", ".rpack", ".grr")
 
     @classmethod
     def is_recognized_file(cls, value):
+        """Return `True` if file type is a recognized type, `False` otherwise"""
         return cls.is_raster_pack_file(value)
 
     @classmethod
     def is_raster_pack_file(cls, value):
+        """Return `True` if file type is GRASS raster pack, `False` otherwise"""
         if isinstance(value, str):
             return value.endswith(cls.raster_pack_suffixes)
         if isinstance(value, Path):
@@ -20,8 +30,11 @@ class ImporterExporter:
     def __init__(self, *, run_function, run_cmd_function):
         self._run_function = run_function
         self._run_cmd_function = run_cmd_function
-        self.input_rasters: list[tuple] = []
-        self.output_rasters: list[tuple] = []
+        # At least for reading purposes, public access to the lists makes sense.
+        self.input_rasters: list[tuple[Path, str]] = []
+        self.output_rasters: list[tuple[Path, str]] = []
+        self.current_input_rasters: list[tuple[Path, str]] = []
+        self.current_output_rasters: list[tuple[Path, str]] = []
 
     def process_parameter_list(self, command, **popen_options):
         """Ingests any file for later imports and exports and replaces arguments
@@ -32,14 +45,25 @@ class ImporterExporter:
         will understand that, i.e., file paths into data names in a project.
         """
         # Get processed parameters to distinguish inputs and outputs.
+        # We actually don't know the type of the input or outputs) because that is
+        # currently not included in --json. Consequently, we are only assuming that the
+        # files are meant to be used as in-project data. So, we need to deal with cases
+        # where that's not true one by one, such as r.unpack taking file,
+        # not raster (cell), so the file needs to be left as is.
         parameters = self._process_parameters(command, **popen_options)
         tool_name = parameters["module"]
         args = command.copy()
+        # We will deal with inputs right away
         if "inputs" in parameters:
             for item in parameters["inputs"]:
                 if tool_name != "r.unpack" and self.is_raster_pack_file(item["value"]):
                     in_project_name = self._to_name(item["value"])
-                    self.input_rasters.append((Path(item["value"]), in_project_name))
+                    record = (Path(item["value"]), in_project_name)
+                    if (
+                        record not in self.input_rasters
+                        and record not in self.current_input_rasters
+                    ):
+                        self.current_input_rasters.append(record)
                     for i, arg in enumerate(args):
                         if arg.startswith(f"{item['param']}="):
                             arg = arg.replace(item["value"], in_project_name)
@@ -48,7 +72,14 @@ class ImporterExporter:
             for item in parameters["outputs"]:
                 if tool_name != "r.pack" and self.is_raster_pack_file(item["value"]):
                     in_project_name = self._to_name(item["value"])
-                    self.output_rasters.append((Path(item["value"]), in_project_name))
+                    record = (Path(item["value"]), in_project_name)
+                    # Following the logic of r.slope.aspect, we don't deal with one output repeated
+                    # more than once, but this would be the place to address it.
+                    if (
+                        record not in self.output_rasters
+                        and record not in self.current_output_rasters
+                    ):
+                        self.current_output_rasters.append(record)
                     for i, arg in enumerate(args):
                         if arg.startswith(f"{item['param']}="):
                             arg = arg.replace(item["value"], in_project_name)
@@ -67,42 +98,68 @@ class ImporterExporter:
     def _to_name(self, value, /):
         return Path(value).stem
 
-    def import_rasters(self):
-        for raster_file, in_project_name in self.input_rasters:
+    def import_rasters(self, rasters, *, env):
+        for raster_file, in_project_name in rasters:
+            # Overwriting here is driven by the run function.
             self._run_function(
                 "r.unpack",
                 input=raster_file,
                 output=in_project_name,
-                overwrite=True,
                 superquiet=True,
+                env=env,
             )
 
-    def export_rasters(self):
+    def export_rasters(
+        self, rasters, *, env, delete_first: bool, overwrite: Literal[True] | None
+    ):
         # Pack the output raster
-        for raster_file, in_project_name in self.output_rasters:
+        for raster_file, in_project_name in rasters:
             # Overwriting a file is a warning, so to avoid it, we delete the file first.
             # This creates a behavior consistent with command line tools.
-            Path(raster_file).unlink(missing_ok=True)
+            if delete_first:
+                Path(raster_file).unlink(missing_ok=True)
 
+            # Overwriting here is driven by the run function and env.
             self._run_function(
                 "r.pack",
                 input=in_project_name,
                 output=raster_file,
                 flags="c",
-                overwrite=True,
                 superquiet=True,
+                env=env,
+                overwrite=overwrite,
             )
 
-    def import_data(self):
-        self.import_rasters()
+    def import_data(self, *, env):
+        # We import the data, make records for later, and the clear the current list.
+        self.import_rasters(self.current_input_rasters, env=env)
+        self.input_rasters.extend(self.current_input_rasters)
+        self.current_input_rasters = []
 
-    def export_data(self):
-        self.export_rasters()
+    def export_data(
+        self, *, env, delete_first: bool = False, overwrite: Literal[True] | None = None
+    ):
+        # We export the data, make records for later, and the clear the current list.
+        self.export_rasters(
+            self.current_output_rasters,
+            env=env,
+            delete_first=delete_first,
+            overwrite=overwrite,
+        )
+        self.output_rasters.extend(self.current_output_rasters)
+        self.current_output_rasters = []
 
-    def cleanup(self):
+    def cleanup(self, *, env):
         remove = [name for (unused, name) in self.input_rasters]
         remove.extend([name for (unused, name) in self.output_rasters])
         if remove:
             self._run_function(
-                "g.remove", type="raster", name=remove, superquiet=True, flags="f"
+                "g.remove",
+                type="raster",
+                name=remove,
+                superquiet=True,
+                flags="f",
+                env=env,
             )
+        self.input_rasters = []
+        self.output_rasters = []
